requirements/requirements_nlpvenv.txt
pandas==2.3.3
pyarrow==22.0.0
kiwipiepy==0.22.0
hanja==0.15.1
boto3==1.41.0
ollama==0.6.1
tqdm==4.67.1
psycopg2-binary==2.9.11
s3fs
google-genai